requirements.txt
flet==0.21.2
google-generativeai>=0.5.0
